Successfully build wheels (#477)
This fixes an issue where the install scripts directory is not created
during the install process (due to the package not containing any
install scripts), but it does rely on the install_directories to
manually install binaries. Fix the manual binary copying logic to check
for the existence of the install_scripts directory and create it if
necessary

python setup.py bdist_wheel should work after this. Additionally, pip
installing the package (even if no wheel exists) will succeed at
generating the wheel dynamically.